AI Bee Debug
Got the AI bee to spawn and finally move to a flower. There was an issue where no matter what it would move to the hive waypoint, despite being focused on a flower. This still needs to be addressed, however for now i simply disabled its ability to move to the hive and it now freezes upon reaching a flower